+++ docs/work-log/답변리팩토링_0429/답변리팩토링_01단계_0429.md
@@ -1,0 +1,52 @@
+# 답변리팩토링\_01단계\_0429
+
+## 단계 요약
+
+- refactor-agent가 답변 등록·수정·채택·댓글 관련 22개 파일을 분석하고 리팩토링 항목을 도출함
+
+## 생성 파일
+
+- 없음 (분석 단계)
+
+## 수정 내역
+
+- 없음 (분석 단계)
+
+## 분석 결과 — 우선순위별 리팩토링 항목
+
+### 높음
+
+| 항목                 | 대상                      | 내용                                                                                                   |
+| -------------------- | ------------------------- | ------------------------------------------------------------------------------------------------------ |
+| barrel export 정합성 | `src/components/index.ts` | qna 미포함 → `export * from './qna'` 추가                                                              |
+| import 경로 일관성   | `QnaDetailPage.tsx:10-15` | `@/components/common/Modal`, `@/components/common/Toast`, `@/components/qna/...` → `@/components` 통일 |
+| import 경로 일관성   | `AnswerForm.tsx:8`        | `@/components/common/Button` → `@/components`                                                          |
+| import 경로 일관성   | `MarkdownEditor.tsx:9`    | `@/components/common/Toast` → `@/components`                                                           |
+| CommentForm 버튼     | `CommentForm.tsx:32-43`   | 원시 button 태그 → 공통 Button 컴포넌트로 교체                                                         |
+| CommentList sortRef  | `CommentList.tsx:14`      | 선언 후 미사용 dead code → 제거                                                                        |
+
+### 중간
+
+| 항목            | 대상                        | 내용                                                       |
+| --------------- | --------------------------- | ---------------------------------------------------------- |
+| AnswerCard 분리 | `QnaDetailPage.tsx:511-601` | 답변 카드 article 블록(~90줄) → AnswerCard 컴포넌트로 분리 |
+
+### 낮음 / 현행 유지
+
+| 항목                                             | 대상                       | 결정                                                            |
+| ------------------------------------------------ | -------------------------- | --------------------------------------------------------------- |
+| LinkIcon, RobotIcon                              | `QnaDetailPage.tsx:32-100` | 이 페이지에서만 사용, 분리 실익 낮음 → 유지                     |
+| MarkdownEditor 인라인 style                      | `MarkdownEditor.tsx:174`   | 서드파티 컴포넌트 prop → 유지                                   |
+| handleCreateSubmit/handleEditSubmit onError 중복 | `QnaDetailPage.tsx`        | 에러 코드별 메시지·action 달라 명시적 유지가 가독성 우수 → 유지 |
+
+## 막혔던 것
+
+- 없음
+
+## 다음 단계 참고
+
+- 작업02(컴포넌트 분리), 작업03(import·스타일 정리)를 분리 실행하되, barrel export 추가를 먼저 해야 import 경로 통일이 가능함
+
+## 커밋
+
+- 미커밋
