+++ docs/work-log/답변리팩토링_0429/답변리팩토링_02단계_0429.md
@@ -1,0 +1,29 @@
+# 답변리팩토링\_02단계\_0429
+
+## 단계 요약
+
+- AnswerCard 컴포넌트 분리, CommentForm 버튼 교체, CommentList sortRef dead code 제거 완료
+
+## 생성 파일
+
+- `src/components/qna/AnswerCard/AnswerCard.tsx`
+- `src/components/qna/AnswerCard/index.ts`
+
+## 수정 내역
+
+- `src/components/qna/CommentList/CommentList.tsx` — `useRef` import, `sortRef` 선언, `ref={sortRef}` 제거
+- `src/components/qna/CommentForm/CommentForm.tsx` — 원시 `<button>` → 공통 `Button` 컴포넌트(`size="sm"`, `loading` prop)로 교체
+- `src/components/qna/index.ts` — `export * from './AnswerCard'` 추가
+- `src/pages/qna/QnaDetailPage.tsx` — `AnswerCard` import 추가, `CommentForm`·`CommentList` 직접 import 제거
+
+## 막혔던 것
+
+- 없음
+
+## 다음 단계 참고
+
+- 작업03(import 정리·barrel export)에서 `src/components/index.ts`에 `export * from './qna'` 추가 및 QnaDetailPage, AnswerForm, MarkdownEditor의 import 경로 통일 필요
+
+## 커밋
+
+- 미커밋
