+++ docs/work-log/답변리팩토링_0429/답변리팩토링_03단계_0429.md
@@ -1,0 +1,29 @@
+# 답변리팩토링\_03단계\_0429
+
+## 단계 요약
+
+- barrel export 정합성 수정 및 import 경로 통일 완료
+
+## 생성 파일
+
+- 없음
+
+## 수정 내역
+
+- `src/components/index.ts` — `export * from './qna'`, `export * from './chatbot'` 추가
+- `src/pages/qna/QnaDetailPage.tsx` — ConfirmModal, Toast, AnswerForm, AnswerCard, AnswerFormHandle을 `@/components` 단일 진입점으로 통일 (6줄 → 2줄)
+- `src/pages/qna/QnaListPage.tsx` — QuestionCard, CategoryFilter를 `@/components`로 통일 (직접 경로 2개 제거)
+- `src/pages/ComponentShowcase.tsx` — `'../components'` → `'@/components'`
+- `src/hooks/useToast.ts` — `@/components/common/Toast` → `@/components`
+
+## 막혔던 것
+
+- 없음 (순환참조 방지 대상 파일은 직접 경로 유지로 명확히 구분)
+
+## 다음 단계 참고
+
+- 작업04(최종 검증)에서 lint·tsc·build 재확인 후 커밋
+
+## 커밋
+
+- 미커밋
