+++ docs/work-log/답변리팩토링_0429/답변리팩토링_0429_계획내용.md
@@ -1,0 +1,21 @@
+# 답변리팩토링*0429*계획내용
+
+## 작업 계획
+
+- 전체 목표: refactor-agent를 사용해 답변 등록·수정·채택·댓글 기능의 코드 구조를 개선한다 (중복 제거, 컴포넌트 분리, 스타일 정합성, import 정리)
+
+## 작업01 — 대상 파일 분석 및 리팩토링 계획 수립
+
+- [x] 완료
+
+## 작업02 — 컴포넌트 분리 실행 (CommentForm, CommentList 등)
+
+- [x] 완료
+
+## 작업03 — 스타일 정합성 및 import 정리
+
+- [x] 완료
+
+## 작업04 — 검증 (lint · tsc · build) 및 결과 정리
+
+- [x] 완료
