+++ docs/work-log/답변리팩토링_0429/답변리팩토링_04단계_0429.md
@@ -1,0 +1,31 @@
+# 답변리팩토링\_04단계\_0429
+
+## 단계 요약
+
+- pnpm lint · npx tsc --noEmit · pnpm build 최종 검증 모두 PASS, 리팩토링 전 과정 완료
+
+## 생성 파일
+
+- 없음
+
+## 수정 내역
+
+- 없음 (검증 단계)
+
+## 막혔던 것
+
+- 없음
+
+## 다음 단계 참고
+
+- 리팩토링 완료. 커밋 후 PR 생성 가능
+
+## 커밋
+
+- 미커밋
+
+## 검증 결과
+
+- lint: PASS
+- 타입 체크 (tsc --noEmit): PASS
+- 빌드 (pnpm build): PASS
