+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export * from './common'
 export * from './layout'
+export * from './qna'
+export * from './chatbot'
