+++ src/components/qna/CommentForm/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentForm'
