+++ src/components/qna/CommentList/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentList'
